backend/requirements.txt
pytest
# LangChain依赖
langchain
langgraph
langchain-community
langchain-core
langchain-openai
langchain-deepseek
langgraph-checkpoint-postgres
asyncpg
openai>=1.0.0  # DeepSeek支持OpenAI SDK格式，但我们使用它来调用DeepSeek API
bcrypt
fastapi
uvicorn
pydantic
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
python-jose
passlib
python-multipart
httpx
# numpy==1.26.2
# 不再需要单独的翻译依赖，使用DeepSeek模型进行翻译
# sentence-transformers>=2.2.0  # 删除，不再需要本地embedding模型
# torch==2.0.1  # 删除，不再需要torch
# docarray==0.32.0  # 删除，与langchain相关
# faiss-gpu  # 删除，不再需要向量搜索
# tavily-python==0.5.1  # 删除，不再需要
# streamlit==1.43.1  # 如果不需要streamlit，可以删除
# langchain-huggingface>=0.0.2  # 删除，不再需要huggingface嵌入

# 保留基础工具依赖
sqlalchemy-utils>=0.36.8  # 数据库工具支持
toml>=0.10.2  # 配置文件支持
typing-extensions>=4.0.0  # 类型支持
pyyaml>=6.0  # YAML配置支持
requests>=2.28.0  # HTTP客户端
pytest-cov>=3.0.0  # 测试覆盖率工具
aiohttp>=3.8.1  # 异步HTTP客户端
# langchain_google_genai  # 删除，不再需要Google AI集成
aiosmtplib
watchfiles
pgvector 
sse-starlette  # 添加 Starlette Server-Sent Events 支持
langchain-google-genai # 添加 Google GenAI 支持 
IPython # Added for visualization and interactive debugging
playwright # Added for LangGraph visualization
pygraphviz # Added for LangGraph visualization 